+++ src/components/RecipeStep/RecipeStepItem.tsx
@@ -1,17 +1,13 @@
 import {Heading, Flex, Box, Textarea, Input} from '@chakra-ui/react';
 import styled from '@emotion/styled';
-import Btn from 'components/Btn';
 import React from 'react';
-
-function RecipeStepItem({
-  step,
-  value,
-  handleStepTextChange,
-}: {
+import ImageFileInput from './ImageFileInput';
+interface RecipeStepItem {
   step: number;
   value: string;
-  handleStepTextChange: any;
-}) {
+  handleStepTextChange: React.ChangeEventHandler<HTMLTextAreaElement>;
+}
+function RecipeStepItem({step, value, handleStepTextChange}: RecipeStepItem) {
   return (
     <>
       <Heading color={'#EA900B'} size="md">
@@ -30,44 +26,16 @@ function RecipeStepItem({
             border={'none'}
             resize="none"
             p={5}
-            // value={value}
+            value={value}
             onChange={handleStepTextChange}
           />
         </Box>
         <Box width={'200px'} border="1px solid #b9b9b9" borderRightRadius={5}>
-          <FileInput step={step} />
+          <ImageFileInput step={step} />
         </Box>
       </Flex>
     </>
   );
 }
-function FileInput({step}: {step: number}) {
-  return (
-    <Flex flexDirection={'column'} justifyContent="center" w="100%" h="100%">
-      <InfoInput type={'file'} id={`recipeImgInput${step}`} hidden />
-      <RecipeImgInputLabel htmlFor={`recipeImgInput${step}`}>
-        <Btn>사진 업로드</Btn>
-      </RecipeImgInputLabel>
-    </Flex>
-  );
-}
-
-const RecipeImgInputLabel = styled.label`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  &::before {
-    content: '요리 사진을 등록해 주세요.';
-  }
-`;
-const InfoInput = styled(Input)`
-  font-size: 1.1rem;
-  width: 100%;
-  height: 100%;
-  padding: 20px;
-  border-top-left-radius: 5px;
-`;
 
 export default RecipeStepItem;
